backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug>=2.3.7
gunicorn==20.1.0
langchain>=0.1.0
langchain-core>=0.1.13
langchain-community>=0.0.13
langchain-text-splitters>=0.3.6
beautifulsoup4==4.12.2
requests>=2.31.0
python-dotenv==1.0.0
pydantic>=1.10.8
aiohttp>=3.7.4
httpx==0.27.2  # Specific version to fix OpenAI client compatibility
# perplexity-ai==0.1.1 - Package not available in PyPI
openai==1.12.0
